Modified login mechanic and ProfileView
